refactor(tests): split fake repository auth
